Move to Reference-based maps and narrow types where possible (#21)
* Move to Reference-based maps and narrow types where possible

* Fix method calls and narrow numeric types in the upgrade config

* Fix compile error

* Fix tricky ClassCastException

+++ src/main/java/me/morpheus/metropolis/api/config/PlotCategory.java
@@ -4,6 +4,6 @@ import me.morpheus.metropolis.api.flag.Flag;
 
 public interface PlotCategory {
 
-    int getUnownedPermission(Flag flag);
+    byte getUnownedPermission(Flag flag);
 
 }
